+++ test/bonus_test.rb
@@ -2,12 +2,12 @@ test 'Bonus'
 
 test 'favicon'
 
-pull '/favicon.ico' rescue @code = 404
+pull '/favicon.ico'
 is @code, 404
 
 
 test 'robots'
 
-pull '/robots.txt' rescue @code = 404
+pull '/robots.txt'
 is @code, 200
-is @body.include?('Sitemap'), true
+is @body.include?('Sitemap')