+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-**Version 0.1.0** - *2017-01-05*
+**Version 0.1.0** - *2017-01-07*
+
+- Options added for robots.txt and favicon.ico
+- Compress and bundle options are working
+- Fixed caching issue
+
+**Version 0.0.3** - *2017-01-05*
 
 - Fixed gemspec issues